5/7/2023 Styling needs to be finished by tmrow
I styled everything that was needed for each page. I added some images for about page like the collage from myself and the 4 year college plan image. Resume had the buttons styled and working for clicking on the pdf. The contact got the buttons styled and the form ready which will be the next step for adding javascript for it. I need to link the pages and work on the writings. html and the projects.html as well as the selfselected samples html styling.